Add job to update PG searchpath if grafana enabled
In order for grafana to auto populate the metric table
names in their UI then the searchpath for the user
used to connect to the db must be updated to include
the prom and _prom_catalog schemas.

The grafana data-source configuration does not
allow the search path to be selected when
configuring or creating the data source.