Configure grafana to use TimescaleDB as a database for itself (#20)
The chart allows a configurable way to create a separate user
and schema that grafana will use for it's data. The same
mechanism is used to create a user that grafana will use for
the TimescaleDB provisioned data source.

Notes for getting grafana admin password are simplified
by being aware whether it's a upgrade or install.