Refactor connector queue config values (#23)
* Refactor timescale-prometheus remote_write configuration

Previously the configuration options for Timescale-Prometheus
were stored under the values for the connector. This created
confusion for users because the prometheus chart has a separate
remoteWrite configuration option. This refactor aims to make
it clear that the configuration for the timescale-prometheus
remote_write will go into the config map override
of prometheus.